requirements.txt
pandas==2.2.1
python-dotenv==1.0.1
httpx==0.27.2
openai==1.56.1
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pydub==0.25.1
requests==2.31.0
ffmpeg-python==0.2.0
numpy==1.26.0
streamlit==1.32.2